Added to Database + one function
Added all of the major scales, and all of the proposed chords, can add more later.
Also added a strict search for finding all of the chords whose notes entirely fall in a scale. It will likely be slow but the only fast way to do it would to be to add yet another database with ScaleName keys and List<ChordName> values, doable but can be done later.